adicion de base de datos de pacientes y busqueda por numero de documento
